test(core): move failing lnd parse-error tests to new structure
